configured apache, added port forwarding to host computer

+++ setup.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-#based on script by chris yip
+
 sudo apt-get -y update
 
 sudo apt-get -y install apache2
@@ -10,10 +10,14 @@ sudo apt-get -y install mysql-server libapache2-mod-auth-mysql php5-mysql
 
 apt-get install -y php5 libapache2-mod-php5
 
-/etc/init.d/apache2 restart
+sed -i '/AllowOverride None/c AllowOverride All' /etc/apache2/sites-available/default
 
-rm -rf /var/www
-ln -fs /vagrant /var/www
+if ! [ -L /var/www ]; then
+	rm -rf /var/www
+	ln -fs /vagrant /var/www
+fi
+
+/etc/init.d/apache2 restart
 
 #set up php unit
 wget https://phar.phpunit.de/phpunit.phar
